flask_survey_app/requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-Security-Too>=5.1.0
Flask-WTF>=1.1.0
WTForms>=3.0.0
Bootstrap-Flask>=2.2.0
Flask-Mail>=0.9.1
email-validator>=2.0.0
bcrypt>=4.0.0
argon2-cffi>=23.1.0
python-dotenv>=1.0.0